% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aggregate_daily.R
\name{aggregate_daily}
\alias{aggregate_daily}
\title{Calculate Daily Sums or Averages}
\usage{
aggregate_daily(time, y, basedate, FUN = sum)
}
\arguments{
\item{time}{vector of POSIXct times}

\item{y}{corresponding vector of data (e.g. hourly)}

\item{basedate}{reference date where the series starts from, e.g.,
"1970-01-01 00:00.00 UTC"}

\item{FUN}{function that is applied to the data of each day}
}
\value{
Data frame with the following columns:
\describe{
  \item{day}{POSIXct time of the days}
  \item{y}{vector of aggregated values}
}
}
\description{
Calculate daily sums or averages, e.g. from from hourly irradiation data.
}
\examples{
## global hourly irradiation data
data(irad)
daily <- aggregate_daily(irad$time, irad$irad2, basedate= "1970-01-01 00:00.00 UTC")
}

