% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/call_salmodll_macrophytes.R
\name{call_salmodll_macrophytes}
\alias{call_salmodll_macrophytes}
\title{Call SALMO Shared Library}
\usage{
call_salmodll_macrophytes(cfunc, nOfVar, cc, pp, uu, xx, pm, mx)
}
\arguments{
\item{cfunc}{string, name of the C function to be called}

\item{nOfVar}{vector with number of variables}

\item{cc}{vector of constants for the model}

\item{pp}{matrix of phytoplankton parameters}

\item{uu}{input vector (environmental conditions)}

\item{xx}{state vector}

\item{pm}{parameters for macrophytes}

\item{mx}{states of macrophytes (ReacTran format)}
}
\value{
list, containing the derivatives as first element
}
\description{
Call the shared library of SALMO (with additional states and parameters for
macrophyte coupling)
}

