% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{data_bautzen_1997}
\alias{data_bautzen_1997}
\title{Input data set data_bautzen_1997}
\description{
Input data set data_bautzen_1997
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}

