% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{hypso_bautzen}
\alias{hypso_bautzen}
\title{Data set hypso_bautzen}
\description{
Hypsographic Table of Bautzen Reservoir
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}

