% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_salmo_integers.R
\name{init_salmo_integers}
\alias{init_salmo_integers}
\title{Initialize Index Values of Parameter List}
\usage{
init_salmo_integers(parms)
}
\arguments{
\item{parms}{list containing constant model parameters}
}
\value{
environment (hash table) that indexes for inputs and states
}
\description{
Add precalculated index values to the parameter list, to avoid repeated
searching.
}
