% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{parms_mac_ctrl}
\alias{parms_mac_ctrl}
\title{parms_mac_ctrl data set}
\description{
parms_mac_ctrl data set
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}

