% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{parms_mac_plants}
\alias{parms_mac_plants}
\title{parms_mac_plants data set}
\description{
parms_mac_plants data set
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}
