% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_data.R
\docType{data}
\name{parms_salmo_ctrl}
\alias{parms_salmo_ctrl}
\title{parms_salmo_ctrl data set}
\description{
parms_salmo_ctrl data set
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}
