% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{parms_salmo_other}
\alias{parms_salmo_other}
\title{parms_salmo_other data set}
\description{
Other SALMO parameters (id, value - pairs)
}
\references{
\url{http://www.tu-dresden.de/hydrobiologie}
}
\keyword{data}

