% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSALMO-package.R
\docType{package}
\name{rSALMO-package}
\alias{rSALMO-package}
\alias{rSALMO}
\title{Ecological Simulation of Lakes}
\description{
\tabular{ll}{
Package:  \tab rSALMO\cr
Type:     \tab Package\cr
Version:  \tab 0.3-1\cr
Date:     \tab 2025-01-25\cr
License:  \tab  GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
An R implementation of the SALMO lake model\cr 
(SALMO: Simulation by means of An analytical Lake MOdel)
}
\examples{

demo("demo_salmo_1box")
demo("demo_salmo_2box")

}
\references{
see \url{http://hhbio.wasser.tu-dresden.de/projects/salmo/}
}
\seealso{
\code{\link[deSolve:deSolve-package]{deSolve}}
}
\author{
Rene Sachse and Thomas Petzoldt (package)\cr

Susanne Rolinski (C code of SALMO)\cr

Juergen Benndorf + (system of equations of SALMO)\cr

Maintainer: Thomas Petzoldt <thomas.petzoldt@tu-dresden.de>\cr
}
\keyword{package}
