% Generated by roxygen2 (4.0.2): do not edit by hand
\name{salmo_1box}
\alias{salmo_1box}
\alias{salmo_2box}
\title{Derivatives of SALMO}
\usage{
salmo_1box(time, x, p, inputs)

salmo_2box(time, x, p, inputs)
}
\arguments{
\item{time}{simulation time (days)}

\item{x}{state vector in correct order}

\item{p}{list containing constant model parameters}

\item{inputs}{input vector (environmental conditions)}
}
\value{
list, first element contains the derivatives, other elements can
contain optional outputs
}
\description{
Main model function to be called from ode solvers (Fully mixed one-box
version).
}

