% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/salmo_mac_box.R
\name{salmo_mac_1box}
\alias{salmo_mac_1box}
\title{Derivatives of SALMO with macrophytes}
\usage{
salmo_mac_1box(time, states, parms, inputs)
}
\arguments{
\item{time}{simulation time (days)}

\item{states}{state vector in correct order}

\item{parms}{list containing constant model parameters}

\item{inputs}{input vector (environmental conditions)}
}
\value{
list, first element contains the derivatives, other elements can
contain optional outputs
}
\description{
Main model function to be called from ode solvers (1 box version with macrophytes).
}

