\docType{package}
\name{simlake-package}
\alias{simlake}
\alias{simlake-package}
\title{Ecological Simulation of Lakes}
\description{
  Ecological Simulation of Lakes
}
\details{
  \tabular{ll}{ Package: \tab simlake\cr Type: \tab
  Package\cr Version: \tab 0.02\cr Date: \tab 2012-04-19\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

  Add a short description of the package .....
}
\examples{
## add a few examples
}
\author{
  Rene Sachse and Thomas Petzoldt (package)\cr

  Susanne Rolinski (C code of SALMO)\cr

  Juergen Benndorf + (system of equations of SALMO)\cr

  Maintainer: Thomas Petzoldt
  <thomas.petzoldt@tu-dresden.de>\cr
}
\references{
  see
  \url{http://hhbio.wasser.tu-dresden.de/projects/salmo/}
}
\seealso{
  \code{\link[deSolve:deSolve-package]{deSolve}}
}
\keyword{package}

