% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeInputVector}
\alias{makeInputVector}
\title{Make Input Vector from List of Input Data}
\usage{
makeInputVector(inputs, x)
}
\arguments{
\item{inputs}{list of input data; must contain vectors depth and time}

\item{time}{vector of requested time steps for interpolation}
}
\value{
vector of interpolated data in the order abcabcabc.
}
\description{
Interpolates consistent vector of input data for a given time point from
a list that contains data of different spatial and temporal resulution.
}

