% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmo_box.R
\name{salmo_1box}
\alias{salmo_1box}
\alias{salmo_2box}
\title{Derivatives of SALMO}
\usage{
salmo_1box(time, x, p, inputs, ndx)

salmo_2box(time, x, p, inputs, ndx)
}
\arguments{
\item{time}{simulation time (days)}

\item{x}{state vector in correct order}

\item{p}{list containing constant model parameters}

\item{inputs}{input vector (environmental conditions)}

\item{ndx}{hashtable (environment) of indexes and counters}
}
\value{
list, first element contains the derivatives, other elements can
contain optional outputs
}
\description{
Main model function to be called from ode solvers (Fully mixed one-box
version).
}
