% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmo_state_names.R
\name{salmo_input_names}
\alias{salmo_input_names}
\alias{salmo_state_names}
\title{Column Names of SALMO State and Input Variables}
\usage{
salmo_input_names()

salmo_state_names(nlayers, macrophytes = FALSE)
}
\arguments{
\item{nlayers}{numeric, number of layers fro replicating the names}

\item{macrophytes}{logical if state variable names for macrophytes should be included}
}
\value{
vector of names for states resp inputs.
contain optional outputs
}
\description{
Column Names of SALMO State and Input Variables
}
