% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setparms}
\alias{setparms}
\title{Set Model Parameters}
\usage{
setparms(parms, values, pnames = NULL, col = 1, slot = NA)
}
\arguments{
\item{parms}{data object containing model parameters as a vector, a matrix
or a non-nested list of vectors and matrices}

\item{values}{vector or list of replacement values}

\item{pnames}{if \code{values} is a vector or matrix:
vector of parameter names that will be replaced}

\item{col}{column in matrix-like parameter objects}

\item{slot}{character, name of a single list element of \code{parms}}
}
\value{
the manipulated \code{parms} object (vector, matrix or list)
}
\description{
Robustly replace model parameters by ignoring capitalisation and order of
replacment.
}
\examples{
data(cc)
cc <- setparms(cc, c("YZP","dummy", "EPSMIN", "dtt"), c(0.7, 66, 0.77, 0.1))
cc

data(pp)
pp <- setparms(pp, c("TOPTX", "vs"), c(28, 0.22), c(1, 3))
# pp <- setparms(pp, "KP", c(1.4, 1.5, 1.6), 1:3) # not yet possible
}

