% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transport.R
\name{transport}
\alias{transport}
\title{Calulate Vertical Transport for a 1D Lake Model}
\usage{
transport(x, forcings, parms, ndx, idzmix, zres, vmat)
}
\arguments{
\item{x}{vector of state variables}

\item{forcings}{matrix of environmental forcing data}

\item{parms}{list of model parameters}

\item{ndx}{hashtable (environment) of indexes and counters}

\item{idzmix}{index of resuspension depth in depth vector}

\item{zres}{resuspension depth}

\item{vmat}{matrix containing the vertical sedimentation velocities for the state variables}
}
\value{
vector of the derivatives of the model
}
\description{
This function calculates vertical transport for a 1D lake model.
It is normally not called by the user.
}

